# src/test/resources/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/return/{bookId}/patron/{patronId}": {
            "put": {
                "tags": [
                    "borrowing-record-controller"
                ],
                "operationId": "returnbook",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "patronId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BorrowingRecordRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BorrowingRecordResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/patrons/{id}": {
            "get": {
                "tags": [
                    "patron-controller"
                ],
                "operationId": "getPatronById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatronResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "patron-controller"
                ],
                "operationId": "updatePatron",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatronRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatronResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "patron-controller"
                ],
                "operationId": "deletePatron",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "enum": [
                                        "100 CONTINUE",
                                        "101 SWITCHING_PROTOCOLS",
                                        "102 PROCESSING",
                                        "103 EARLY_HINTS",
                                        "103 CHECKPOINT",
                                        "200 OK",
                                        "201 CREATED",
                                        "202 ACCEPTED",
                                        "203 NON_AUTHORITATIVE_INFORMATION",
                                        "204 NO_CONTENT",
                                        "205 RESET_CONTENT",
                                        "206 PARTIAL_CONTENT",
                                        "207 MULTI_STATUS",
                                        "208 ALREADY_REPORTED",
                                        "226 IM_USED",
                                        "300 MULTIPLE_CHOICES",
                                        "301 MOVED_PERMANENTLY",
                                        "302 FOUND",
                                        "302 MOVED_TEMPORARILY",
                                        "303 SEE_OTHER",
                                        "304 NOT_MODIFIED",
                                        "305 USE_PROXY",
                                        "307 TEMPORARY_REDIRECT",
                                        "308 PERMANENT_REDIRECT",
                                        "400 BAD_REQUEST",
                                        "401 UNAUTHORIZED",
                                        "402 PAYMENT_REQUIRED",
                                        "403 FORBIDDEN",
                                        "404 NOT_FOUND",
                                        "405 METHOD_NOT_ALLOWED",
                                        "406 NOT_ACCEPTABLE",
                                        "407 PROXY_AUTHENTICATION_REQUIRED",
                                        "408 REQUEST_TIMEOUT",
                                        "409 CONFLICT",
                                        "410 GONE",
                                        "411 LENGTH_REQUIRED",
                                        "412 PRECONDITION_FAILED",
                                        "413 PAYLOAD_TOO_LARGE",
                                        "413 REQUEST_ENTITY_TOO_LARGE",
                                        "414 URI_TOO_LONG",
                                        "414 REQUEST_URI_TOO_LONG",
                                        "415 UNSUPPORTED_MEDIA_TYPE",
                                        "416 REQUESTED_RANGE_NOT_SATISFIABLE",
                                        "417 EXPECTATION_FAILED",
                                        "418 I_AM_A_TEAPOT",
                                        "419 INSUFFICIENT_SPACE_ON_RESOURCE",
                                        "420 METHOD_FAILURE",
                                        "421 DESTINATION_LOCKED",
                                        "422 UNPROCESSABLE_ENTITY",
                                        "423 LOCKED",
                                        "424 FAILED_DEPENDENCY",
                                        "425 TOO_EARLY",
                                        "426 UPGRADE_REQUIRED",
                                        "428 PRECONDITION_REQUIRED",
                                        "429 TOO_MANY_REQUESTS",
                                        "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
                                        "451 UNAVAILABLE_FOR_LEGAL_REASONS",
                                        "500 INTERNAL_SERVER_ERROR",
                                        "501 NOT_IMPLEMENTED",
                                        "502 BAD_GATEWAY",
                                        "503 SERVICE_UNAVAILABLE",
                                        "504 GATEWAY_TIMEOUT",
                                        "505 HTTP_VERSION_NOT_SUPPORTED",
                                        "506 VARIANT_ALSO_NEGOTIATES",
                                        "507 INSUFFICIENT_STORAGE",
                                        "508 LOOP_DETECTED",
                                        "509 BANDWIDTH_LIMIT_EXCEEDED",
                                        "510 NOT_EXTENDED",
                                        "511 NETWORK_AUTHENTICATION_REQUIRED"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "getBookById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/patrons": {
            "get": {
                "tags": [
                    "patron-controller"
                ],
                "operationId": "getAllPatrons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PatronResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "patron-controller"
                ],
                "operationId": "createPatron",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatronRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatronResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/borrow/{bookId}/patron/{patronId}": {
            "post": {
                "tags": [
                    "borrowing-record-controller"
                ],
                "operationId": "borrowbook",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "patronId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BorrowingRecordRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BorrowingRecordResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "getAllBooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookResponseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "createBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BorrowingRecordRequestDTO": {
                "type": "object",
                "properties": {
                    "borrowDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "returnDate": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "required": [
                    "borrowDate",
                    "returnDate"
                ]
            },
            "BorrowingRecordResponseDTO": {
                "type": "object",
                "properties": {
                    "borrowDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "returnDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "bookId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "patronId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "bookName": {
                        "type": "string"
                    },
                    "patronName": {
                        "type": "string"
                    }
                }
            },
            "PatronRequestDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "name"
                ]
            },
            "PatronResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    }
                }
            },
            "BookRequestDTO": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1
                    },
                    "author": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1
                    },
                    "publicationYear": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isbn": {
                        "type": "string",
                        "pattern": "^(\\d{13})$"
                    }
                },
                "required": [
                    "author",
                    "title"
                ]
            },
            "BookResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "author": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "publicationYear": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            }
        }
    }
}